add handler errors

+++ src/main/java/com/foro/domain/topic/TopicService.java
@@ -1,12 +1,14 @@
 package com.foro.domain.topic;
 
 import com.foro.domain.course.CourseRepository;
+import com.foro.domain.exceptios.RequestException;
 import com.foro.domain.user.UserRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -26,17 +28,21 @@ public class TopicService {
 
 
     public Topic createTopic( RegisterTopicDto topicDto){
+
+        if(topicRepository.findByTitleAndMessage(topicDto.title(),topicDto.message()).isPresent()){
+            throw new RequestException(" topic already exist",HttpStatus.BAD_REQUEST);
+        }
          Topic topic =new Topic();
 
 
         var author=userRepository.findById(topicDto.idAuthor());
         if(author.isEmpty()){
-            throw new RuntimeException("author doesn't exist with this id");
+            throw new RequestException("author doesn't exist with this id " + topicDto.idAuthor(),HttpStatus.BAD_REQUEST);
         }
         author.ifPresent(topic::setAuthor);
         var course= courseRepository.findById(topicDto.idCourse());
         if(course.isEmpty()){
-            throw  new RuntimeException("course  doesn't exist with this id");
+            throw  new RequestException("course  doesn't exist with this id "+topicDto.idCourse(),HttpStatus.BAD_REQUEST);
         }
 
          course.ifPresent(topic::setCourse);
@@ -63,7 +69,7 @@ public class TopicService {
     public  Topic getTopic(Long id){
         var topic= topicRepository.findById(id);
         if(topic.isEmpty()){
-            throw new RuntimeException("doesn't topic exist with "+id +" id");
+            throw new RequestException("doesn't topic exist with "+id +" id", HttpStatus.BAD_REQUEST);
         }
         return topic.get();
     }
@@ -72,12 +78,12 @@ public class TopicService {
         var topic=getTopic(id);
         var author=userRepository.findById(registerTopicDto.idAuthor());
         if(author.isEmpty()){
-            throw new RuntimeException("author doesn't exist with this id");
+            throw new RequestException("author doesn't exist with this id " + registerTopicDto.idAuthor(),HttpStatus.BAD_REQUEST);
         }
         author.ifPresent(topic::setAuthor);
         var course= courseRepository.findById(registerTopicDto.idCourse());
         if(course.isEmpty()){
-            throw  new RuntimeException("course  doesn't exist with this id");
+            throw  new RequestException("course  doesn't exist with this id "+registerTopicDto.idCourse(),HttpStatus.BAD_REQUEST);
         }
 
         course.ifPresent(topic::setCourse);
@@ -90,7 +96,9 @@ public class TopicService {
     }
 
     public  ResponseEntity deleteTopic(Long id){
-
+        if(!topicRepository.existsById(id)){
+            throw new RequestException ("topic doesn't exist with this id "+id,HttpStatus.BAD_REQUEST);
+        }
         topicRepository.deleteById(id);
         return ResponseEntity.status(HttpStatusCode.valueOf(204)).build();
     }
@@ -99,8 +107,5 @@ public class TopicService {
 
 
 
-    public Boolean tipicIsExist(String title, String message){
-        var topic= topicRepository.findByTitleAndMessage(title,message);
-        return topic.isPresent();
-    }
+
 }
